Reading Input Files

Capacity Reserve Margin

GenX.load_cap_reserve_margin — Method

load_cap_reserve_margin(setup::Dict, path::AbstractString, sep::AbstractString, inputs_crm::Dict, network_var::DataFrame)

Function for reading input parameters related to planning reserve margin constraints

$CO_2$ Emissions Cap

GenX.load_co2_cap — Method

load_co2_cap(setup::Dict, path::AbstractString, sep::AbstractString, inputs_co2::Dict)

Function for reading input parameters related to CO$_2$ emissions cap constraints

Energy Share Requirement

GenX.load_energy_share_requirement — Method

load_energy_share_requirement(setup::Dict, path::AbstractString, sep::AbstractString, inputs_ESR::Dict)

Function for reading input parameters related to mimimum energy share requirement constraints (e.g. renewable portfolio standard or clean electricity standard policies)

Fuels Data

GenX.load_fuels_data — Method

load_fuels_data(setup::Dict, path::AbstractString, sep::AbstractString, inputs_fuel::Dict)

Function for reading input parameters related to fuel costs and CO$_2$ content of fuels

Generators Input Data

GenX.load_generators_data — Method

load_generators_data(setup::Dict, path::AbstractString, sep::AbstractString, inputs_gen::Dict, fuel_costs::Dict, fuel_CO2::Dict)

Function for reading input parameters related to electricity generators (plus storage and flexible demand resources)

Variability of Generators' Outputs

GenX.load_generators_variability — Method

load_generators_variability(setup::Dict, path::AbstractString, sep::AbstractString, inputs_genvar::Dict)

Function for reading input parameters related to hourly maximum capacity factors for all generators (plus storage and flexible demand resources)

Load Data

GenX.load_load_data — Method

load_data(setup::Dict, path::AbstractString, sep::AbstractString, inputs_load::Dict)

Function for reading input parameters related to electricity load (demand)

Minimum Capacity Requirements

GenX.load_minimum_capacity_requirement — Method

load_minimum_capacity_requirement(path::AbstractString,sep::AbstractString, inputs::Dict, setup::Dict)

Function for reading input parameters related to mimimum capacity requirement constraints (e.g. technology specific deployment mandates)

Transmission Network

GenX.load_network_data — Method

load_network_data(setup::Dict, path::AbstractString, sep::AbstractString, inputs_nw::Dict)

Function for reading input parameters related to the electricity transmission network

Mapping Representative Time Periods

GenX.load_period_map — Method

load_period_map(setup::Dict,path::AbstractString,sep::AbstractString, inputs::Dict)

Function for reading input parameters related to mapping of representative time periods to full chronological time series

Reserves Data

GenX.load_reserves — Method

load_reserves(setup::Dict,path::AbstractString,sep::AbstractString, inputs_res::Dict)

Function for reading input parameters related to frequency regulation and operating reserve requirements