Functions for Writing the Different Results/Outputs to Separate Files

GenX.write_outputs — Method

write_outputs(EP::Model, path::AbstractString, setup::Dict, inputs::Dict)

Function for the entry-point for writing the different output files. From here, onward several other functions are called, each for writing specific output files, like costs, capacities, etc.

Write Status

GenX.write_status — Method

write_status(path::AbstractString, sep::AbstractString, inputs::Dict, EP::Model)

Function for writing the final solve status of the optimization problem solved

Write Costs

GenX.write_costs — Method

write_costs(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for writing the costs pertaining to the objective function (fixed, variable O & M etc.)

Write Capacities

GenX.write_capacity — Method

write_capacity(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model))

Function for writing the diferent capacities for the different generation technologies (starting capacities or, existing capacities, retired capacities, and new-built capacities)

Write Charge Values

GenX.write_charge — Method

write_charge(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for writing the charging energy values of the different storage technologies.

Write Storage

GenX.write_storage — Method

write_storage(path::AbstractString, sep::AbstractString, inputs::Dict,setup::Dict, EP::Model)

Function for writing the capacities of different storage technologies, including hydro reservoir, flexible storage tech etc.

Write Power

GenX.write_power — Method

write_power(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for writing the different values of power generated by the different technologies in operation

Write Curtailment

GenX.write_curtailment — Method

write_curtailment(path::AbstractString, sep::AbstractString, inputs::Dict, setup::Dict, EP::Model)

Function for writing the curtailment values of the different variable renewable resources.