Investment Energy

GenX.investment_energy! — Method

investment_energy!(EP::Model, inputs::Dict)

This function defines the expressions and constraints keeping track of total available storage charge capacity across all resources as well as constraints on capacity retirements. The function also adds investment and fixed O\&M related costs related to charge capacity to the objective function.

The total capacity of each resource is defined as the sum of the existing capacity plus the newly invested capacity minus any retired capacity.

\[\begin{aligned} & \Delta^{total,energy}_{y,z} =(\overline{\Delta^{energy}_{y,z}}+\Omega^{energy}_{y,z}-\Delta^{energy}_{y,z}) \forall y \in \mathcal{O}, z \in \mathcal{Z} \end{aligned}\]

One cannot retire more capacity than existing capacity.

\[\begin{aligned} &\Delta^{energy}_{y,z} \leq \overline{\Delta^{energy}_{y,z}} \hspace{4 cm} \forall y \in \mathcal{O}, z \in \mathcal{Z} \end{aligned}\]

For resources where $\overline{\Omega_{y,z}^{energy}}$ and $\underline{\Omega_{y,z}^{energy}}$ is defined, then we impose constraints on minimum and maximum power capacity.

\[\begin{aligned} & \Delta^{total,energy}_{y,z} \leq \overline{\Omega}^{energy}_{y,z} \hspace{4 cm} \forall y \in \mathcal{O}, z \in \mathcal{Z} \\ & \Delta^{total,energy}_{y,z} \geq \underline{\Omega}^{energy}_{y,z} \hspace{4 cm} \forall y \in \mathcal{O}, z \in \mathcal{Z} \end{aligned}\]

In addition, this function adds investment and fixed O\&M related costs related to charge capacity to the objective function:

\[\begin{aligned} & \sum_{y \in \mathcal{O} } \sum_{z \in \mathcal{Z}} \left( (\pi^{INVEST,energy}_{y,z} \times \Omega^{energy}_{y,z}) + (\pi^{FOM,energy}_{y,z} \times \Delta^{total,energy}_{y,z})\right) \end{aligned}\]